requirements.txt
fastapi==0.98.0
uvicorn==0.22.0
sqlalchemy>=2.0.0
alembic==1.12.0
python-multipart>=0.0.6
boto3==1.28.5
pydantic==1.10.10
pydantic[email]
python-dotenv==1.0.0
python-jose>=3.3.0
passlib>=1.7.4
Flask==2.0.1
mongoengine==0.23.1
pymongo==4.4.0
motor==3.2.0
email-validator==2.0.0